requirements.txt
flask
flask-sqlalchemy
flask-limiter
openai
python-dotenv